Allow custom field name for hashed password
This allows altering the (currently hardcoded) hashed_password field
name.

I thought about perhaps allowing a custom value to be passed into
`hash_password` but it seems like this is something that should be
configured at compile time. Also, I guess it's very much an edge case
for this to be re-used in the same codebase with a different hashed
password field.